Rewrite to use Javalin as server.
This commit changes the server implementation from Spark to Javalin, as Spark is no longer maintained. Many concepts are similar and so there are only some changes in the routes.

Notable difference is that the clojure for a route must write the result directly (it is a consumer of a context) in contrast to being a function as it is in Spark.

Also the pom has been updated to use the latest driver and build plugin as well as a defined encoding.